// src/enums/consultation.ts
export enum STATUS {
  DRAFT = 'draft',
  PENDING = 'pending',
  PRESCRIBED = 'prescribed',
  PROCESSING = 'processing',
  ACCEPTED = 'accepted',
  REJECTED = 'rejected',
}
export enum CONSULTATION_TYPE {
  REGULAR = 'regular',
  VIDEO = 'video',
  FORWARDTO = 'forwardto',
  MEDICATION = 'medication',
}